refactor: action과 reducer를 분리한다
