+++ src/actions.js
@@ -1,0 +1,6 @@
+const updateField = ({ field, value }) => ({
+  type: 'UPDATE_FIELD',
+  payload: { field, value },
+});
+
+export default updateField;
